requirements.txt
Django==4.0.1
djangorestframework==3.13.1
psycopg2==2.9.2
requests==2.27.1
pytest-django==4.5.2
gunicorn==20.1.0
django-debug-toolbar==3.5.0
